requirements.txt
spotipy==2.23.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
folium==0.14.0
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
jupyter==1.0.0
openpyxl==3.1.2